Added effective_genome_size comments
Added a subroutine to calculate the effective genome size of a selected genome.  This value may be used by tools such as the Nextflow ChIP-seq pipeline.  This subroutine has been commented out, but may be incorporated at a later date.  At present pipeline are caculating this value on the fly (which I think is actually preferable - despite the pipeline generating a warining message).